MOW-1364: i believe this is logically sound but the tests fail

+++ common/src/main/java/org/geogebra/common/euclidian/DrawableList.java
@@ -35,7 +35,7 @@ public class DrawableList extends ArrayList<Drawable> {
 		this.comparator = new Comparator<Drawable>() {
 			@Override
 			public int compare(Drawable a, Drawable b) {
-				return comparator.compare(a.geo, b.geo, false);
+				return (int) comparator.compare(a.geo, b.geo, false);
 			}
 		};
 	}
