+++ common/src/main/java/org/geogebra/common/kernel/geos/GeoPriorityComparator.java
@@ -11,5 +11,7 @@ public interface GeoPriorityComparator {
 	 * @return negative if `a` has a higher priority, positive, if `b` has a higher
 	 *            priority, 0, if their priorities are equal
 	 */
-	int compare(GeoElement a, GeoElement b, boolean checkLastHitType);
+
+	float compare(GeoElement a, GeoElement b, boolean checkLastHitType);
+
 }
